move utils into schedule component
